Working on separating tasks to different classes
