+++ lib/tlsobs_scan.py
@@ -12,22 +12,22 @@ coloredlogs.install(level='INFO', logger=logger, reconfigure=True,
                     datefmt="%Y-%m-%d %I:%M:%S %p %Z")
 
 
-class SSHScanTask(task):
-    def __init__(self, target_obj, port=22):
+class MozillaTLSObservatoryTask(task):
+
+    def __init__(self, target_obj):
         super().__init__(target_obj)
-        self.port = port
 
     def run(self):
-        if find_executable('ssh_scan'):
+        if find_executable('tlsobs'):
             # Found in path, run the command
-            # logger.info("[+] Running ssh_scan...")
-            cmd = "ssh_scan -t " + self.task_target.targetdomain + " -p " \
-                + str(self.ssh_port) + " -o /app/results/" + self.tasktarget.targetdomain \
-                + "/ssh_scan.txt"
-            sshscan_cmd = utils.sanitise_shell_command(cmd)
-            p = subprocess.Popen(sshscan_cmd, stdout=subprocess.PIPE, shell=True)
+            logger.info("[+] Running TLS Observatory scan...")
+            cmd = "tlsobs -r -raw " + self.tasktarget.targetname \
+                + " > /app/results/" + self.tasktarget.targetdomain \
+                + "/tlsobs_scan.txt"
+            tlsobs_cmd = utils.sanitise_shell_command(cmd)
+            p = subprocess.Popen(tlsobs_cmd, stdout=subprocess.PIPE, shell=True)
             p.wait()
             return p
         else:
-            logger.error("[-] ssh_scan not found!")
+            logger.error("[-] TLS Observatory not found in Docker image!")
             return False
